ClientAndSuppliers

+++ ERP_System.BLL/People/UserBll.cs
@@ -231,14 +231,7 @@ namespace ERP_System.BLL
                     user.PasswordHash = userDTO.Password.EncryptString();
                 }
 
-                if (userDTO.UserTypeId == Guid.Parse("8484E624-C5A0-463E-986A-66A118D1F2EB"))
-                    user.UserClassification = UserClassification.Client;
-                else if (userDTO.UserTypeId == Guid.Parse("DF5D2D3C-655A-431E-93A3-AC4AF07C8805"))
-                    user.UserClassification = UserClassification.Suppliers;
-
-
-
-
+                
                 user.Salt = AppConstants.EncryptKey;
 
                 var AllStock = new List<UserStock>();
@@ -280,16 +273,13 @@ namespace ERP_System.BLL
                 }
 
                 tbl.UserName = user.UserName;
+                tbl.Name = user.Name;
                 tbl.Email = user.Email;
                 tbl.IsActive = user.IsActive;
                 tbl.UserTypeId = user.UserTypeId;
                 tbl.AddedBy = tbl.AddedBy;
 
-                if (user.UserTypeId == Guid.Parse("8484E624-C5A0-463E-986A-66A118D1F2EB"))
-                    tbl.UserClassification = UserClassification.Client;
-                else if (user.UserTypeId == Guid.Parse("DF5D2D3C-655A-431E-93A3-AC4AF07C8805"))
-                    tbl.UserClassification = UserClassification.Suppliers;
-
+                
                 var AllStock = new List<UserStock>();
 
                 if (userDTO.StockIds != null && userDTO.StockIds.Count() > 0)
@@ -309,8 +299,11 @@ namespace ERP_System.BLL
 
                 if (_repoUser.Update(tbl))
                 {
-                    _repoUserStock.ExecuteStoredProcedure<int>($"DELETE FROM [People].[UserStocks] us WHERE us.UserId='${tbl.ID}'", null, CommandType.StoredProcedure);
-                    _repoUserStock.InsertRange(AllStock);
+                    if (user.UserClassification != UserClassification.Suppliers  && user.UserClassification != UserClassification.Client)
+                    {
+                        _repoUserStock.ExecuteStoredProcedure<int>($"DELETE FROM [People].[UserStocks] us WHERE us.UserId='${tbl.ID}'", null, CommandType.StoredProcedure);
+                        _repoUserStock.InsertRange(AllStock);
+                    }
 
                     resultViewModel.Status = true;
                     resultViewModel.Message = AppConstants.Messages.SavedSuccess;
@@ -367,8 +360,9 @@ namespace ERP_System.BLL
         }
 
         #region LoadData
-        public DataTableResponse LoadData(DataTableRequest mdl)
+        public DataTableResponse LoadData(DataTableRequest mdl, UserClassification? classification )
         {
+            mdl.UserClassification = classification;
             var data = _repoUser.ExecuteStoredProcedure<UserTableDTO>
                 (_spUsers, mdl?.ToSqlParameter(), CommandType.StoredProcedure);
